feat: seperate strategy from policy
