--- simaple/simulate/policy/__init__.py
@@ -1,5 +1,0 @@
-from simaple.simulate.policy.operation import get_operations
-
-__all__ = [
-    "get_operations",
-]
