+++ simaple/simulate/strategy/__init__.py
@@ -1,0 +1,5 @@
+from simaple.simulate.strategy.operation import get_operations
+
+__all__ = [
+    "get_operations",
+]
