+++ simaple/simulate/strategy/dsl.py
@@ -1,7 +1,7 @@
 import functools
 from typing import Callable, Generator
 
-from simaple.simulate.policy.base import OperationGeneratorProto, PolicyContextType
+from simaple.simulate.strategy.base import OperationGeneratorProto, PolicyContextType
 from simaple.simulate.policy.parser import parse_dsl_to_operations
 
 DSLGenerator = Generator[str, PolicyContextType, PolicyContextType]
